add new method

+++ phoenix-knowledge-api/src/main/java/com/ginkgocap/ywxt/knowledge/service/common/KnowledgeBaseService.java
@@ -7,5 +7,5 @@ import com.ginkgocap.ywxt.knowledge.utils.KnowledgeConstant;
  */
 public interface KnowledgeBaseService {
     public final long APPID = KnowledgeConstant.DEFAULT_APP_ID;
-    public final int sourceType = KnowledgeConstant.DEFAULT_SOURCE_TYPE;
+    public final short sourceType = KnowledgeConstant.DEFAULT_SOURCE_TYPE;
 }
