+++ phoenix-knowledge-api/src/main/java/com/ginkgocap/ywxt/knowledge/utils/KnowledgeConstant.java
@@ -12,7 +12,7 @@ import java.util.List;
 public class KnowledgeConstant {
 
     public static final long DEFAULT_APP_ID = 1L;
-    public static final int DEFAULT_SOURCE_TYPE = 8;
+    public static final short DEFAULT_SOURCE_TYPE = 8;
 
     public static final String USER_DEFAULT_PIC_PATH_FAMALE = "/web/pic/user/default.jpeg";//个人默认头像：女
     public static final String USER_DEFAULT_PIC_PATH_MALE = "/web/pic/people/original/default.jpeg";//个人默认头像：男
